+++ spec/lib/castle/extractors/client_id_spec.rb
@@ -3,7 +3,8 @@
 describe Castle::Extractors::ClientId do
   subject(:extractor) { described_class.new(request, cookies) }
 
-  let(:client_id) { 'abcd' }
+  let(:client_id_cookie) { 'abcd' }
+  let(:client_id_header) { 'abcde' }
   let(:cookies) { request.cookies }
   let(:request) { Rack::Request.new(env) }
   let(:env) do
@@ -14,12 +15,12 @@ describe Castle::Extractors::ClientId do
     let(:headers) do
       {
         'HTTP_X_FORWARDED_FOR' => '1.2.3.4',
-        'HTTP_COOKIE' => "__cid=#{client_id};other=efgh"
+        'HTTP_COOKIE' => "__cid=#{client_id_cookie};other=efgh"
       }
     end
 
     it do
-      expect(extractor.call('__cid')).to eql(client_id)
+      expect(extractor.call).to eql(client_id_cookie)
     end
   end
 
@@ -27,12 +28,12 @@ describe Castle::Extractors::ClientId do
     let(:headers) do
       {
         'HTTP_X_FORWARDED_FOR' => '1.2.3.4',
-        'HTTP_X_CASTLE_CLIENT_ID' => client_id
+        'HTTP_X_CASTLE_CLIENT_ID' => client_id_header
       }
     end
 
     it 'appends the client_id' do
-      expect(extractor.call('__cid')).to eql(client_id)
+      expect(extractor.call).to eql(client_id_header)
     end
   end
 
@@ -41,7 +42,21 @@ describe Castle::Extractors::ClientId do
     let(:headers) { {} }
 
     it do
-      expect(extractor.call('__cid')).to eql('')
+      expect(extractor.call).to eql('')
+    end
+  end
+
+  context 'with X-Castle-Client-Id header and cookies client' do
+    let(:headers) do
+      {
+        'HTTP_X_FORWARDED_FOR' => '1.2.3.4',
+        'HTTP_X_CASTLE_CLIENT_ID' => client_id_header,
+        'HTTP_COOKIE' => "__cid=#{client_id_cookie};other=efgh"
+      }
+    end
+
+    it 'appends the client_id' do
+      expect(extractor.call).to eql(client_id_header)
     end
   end
 end